+++ changelog.md
@@ -1,3 +1,7 @@
+# Next
+
+* Add `useInstalledVersions` option
+
 # 1.0.1
 
 * Thanks to @helloIAmPau for adding an extra check when defining deliberate packages in `basePackageValues`.